feat: add a config to generate an op-deployer upgrade to v2.0.0
NOTE - this is checked into the "boba-holocene" branch for convenience,
but it is intended to run under "op-contracts/v2.0.0". The absolute
prestate should also be updated if the corresponding binaries are updated.

The op-deployer upgrade command generates calldata for upgrade() in a
version of OPContractsManager deployed to match the boba-sepolia
Superchain config. It is at 0x99a546a50f0ccd9d2936793b78e378b0f5a83b92.
The transaction must be initiated via delegatecall() from the multisig
for our ProxyAdminOwner.

Changes to be committed:
	new file:   boba-sepolia/upgrade-v2.0.0.json